requirements.txt
nltk==3.5
urllib3==1.25.9
requests==2.24.0
regex==2020.6.8
pdfminer-six==20201018